Fix broken classpath errors
Errors were being caused by the lack of some classes on the server side.
For example, the PayaraArquillianRunner would compile, but wouldn't run
server side tests since the class isn't present on the server. This
commit adds a utility class to package all of these classes in.

This change also removes the use of the CharUtils Google Guava class,
which similarly could not be guaranteed to be present on the server.
Interestingly this would run fine on older versions, but Guava has
recently been removed in Payara 5.193-SNAPSHOT.